Add gitignore for NPM
An NPM-based build process (https://www.npmjs.com) generates some files
that are really annoying if they're saved in the Git repository. Adding
these files to a .gitignore file forces Git to ignore them.

This commit adds the following to the .gitignore:

node_modules/ - Contains all the code for the dependencies installed by
NPM, and their dependencies, and their dependencies, and so on. Even if
the project only needs a few dependencies, this directory can quickly
become huge, and inflate the overall filesize of the repository. Because
everything here can quickly be reinstalled with a single command - npm
install - there's no reason to include in in version control.

package-lock.json - A file that locks the versions of this project's NPM
dependencies. It's supposed to be used for easily rolling back a
repository to see which version of which dependency introduced a bug,
and it's generally recommended that it should be committed. However,
I've never used it, and I generally find that it's more of a nuisance
than anything else.

dist/ - This is the directory that contains the output of the build
process: the compiled, minified CSS files. These files belong in the
gh-pages branch - they should never be committed to the main branch,
which is why the directory is marked as ignored.