Move contents of docs/ to src/
docs/ was a GitHub-specific folder for distributing content via GitHub
Pages, that are intended for distribution of documentation. We'll be
converting the deployment process to an automatic build, so this naming
scheme is no longer required. This commit effectively switches to more
standard directory naming conventions by having the source files be in
the src/ dir.

Example of these conventions:
https://github.com/kriasoft/Folder-Structure-Conventions