Move old minified versions to legacy/
There were some old minified CSS files in src/min/. There are not
supposed to be any minified source files - minification is for
distribution files - so these had to be removed. However, I'm not sure
which if any of them are still being used, and as myself and Grombald
discussed, it's not possible to dynamically generate them using Git
versioning because they were created before the codebase moved to Git.

Instead, I've moved them to a legacy/ dir, and I will configure the task
runner to move these files directly to the minified distribution dir
when the project is built.

New files should not be created in the legacy/ dir.