Set up Gulp task runner
This commit adds Gulp, a task runner, to build the CSS.

If `npm run build` is run on the command line, Gulp will build the
project by passing all CSS through PostCSS and moving all the files to
dist/. The dist/ directory is in the .gitignore, so developers building
the project on their local machine will not commit it to the repository.

Later, I will set up a GitHub Actions workflow that builds the project
on every commit to the main branch. Instead of committing the dist/
directory to the main branch, it will be committed to the gh-pages
branch.

PostCSS is not currently configured to do anything.